Device53 VS Удаление предупреждений

+++ sources/S8-53/src/Menu/Pages/PageDebug.cpp
@@ -485,8 +485,8 @@ static void OnChanged_ADC_Balance_Mode(bool active)
 {
     Draw_ADC_Balance_Mode(0, 0);
 
-    FPGA::WriteToHardware(WR_ADD_RSHIFT_DAC1, shiftADCA, false);
-    FPGA::WriteToHardware(WR_ADD_RSHIFT_DAC2, shiftADCB, false);
+    FPGA::WriteToHardware(WR_ADD_RSHIFT_DAC1, (uint8)shiftADCA, false);
+    FPGA::WriteToHardware(WR_ADD_RSHIFT_DAC2, (uint8)shiftADCB, false);
 }
 
 static void Draw_ADC_Balance_Mode(int x, int y)
@@ -514,7 +514,7 @@ static const Governor mgADC_Balance_ShiftA
 static void OnChanged_ADC_Balance_ShiftA(void)
 {
     BALANCE_ADC_A = shiftADCA;
-    FPGA::WriteToHardware(WR_ADD_RSHIFT_DAC1, BALANCE_ADC_A, false);
+    FPGA::WriteToHardware(WR_ADD_RSHIFT_DAC1, (uint8)BALANCE_ADC_A, false);
 }
 
 static bool IsActive_ADC_Balance_Shift(void)
@@ -535,7 +535,7 @@ static const Governor mgADC_Balance_ShiftB
 static void OnChanged_ADC_Balance_ShiftB(void)
 {
     BALANCE_ADC_B = shiftADCB;
-    FPGA::WriteToHardware(WR_ADD_RSHIFT_DAC2, BALANCE_ADC_B, false);
+    FPGA::WriteToHardware(WR_ADD_RSHIFT_DAC2, (uint8)BALANCE_ADC_B, false);
 }
